ai_service/requirements.txt
fastapi
uvicorn
python-dotenv
groq
langchain
langchain-groq
langchain_community
langchain_core
requests
